Posting Sample Calculations
This commit pushes a set of caclulations for a pellet impact that lands at location (10,20).

The caclulations start with an estimate of 28.81 mm and converges to the expected value of 31.62 mm

Start by entering an estimate of 28.81 (= (70.71-13.10)/2.  This represents the smallest location that will yield a real value for the law of cosints.

Copy the value of 30.84 to the estimate, and the solution should converge in 4 cycles.