Adding diagnostics field to the JSON score
In addition to the score that is transmitted by the target, it was found that there was a need to transmit diagnostics information for debugging and calibration.

The new fields are
North, East, South, and West, the timers values from the microphones
Reference Voltage, the trip point
Temperature, current air temperature.

These fields can be ignored when displaying the shot information.