Additons to parse target JSON message
When connecting to the Target Arduino, it is necessary to enagle DTR on the remore board.  This was done on line 161.

The other change for the Arduino was to accumulate theJSON messages in a buffer and processs the buffer when both a { and } are present.  Once done, the processed message is stripped off and the program continues until the JSON payload is empty.

A third change was to discard a JSON message that has the string 'nan' (Not A Number) . This is an indication that something went wrong in the target caclulations and the payload should be discarded.