Adding in sensor angle compensation
The sensors are ideally located at the compass cardinal points.

It was found that the fragments of target were accumulating on the microphone and would result in a bad reading.

This patch allows the sensors to be located somewhere else (ex 45 degree offset) and then the calculations adjusted to reflect the physical location.