Corrections to the sensor position location
The position of the sensors can introduce a small error in the calculation of the shot location.  Unfortunatly decimal targets rely on small increments, so a small error in the sensor position does yield an error in the resulting target.

This patch allows for a small adjustment to the sensor position to be added to the calculations.