Schematic  Revised to 2.2B
The LM339 was referenced to ground.  In practice the sound signal could go negative as much as 1.3VDC.  The absolute maximum rating for this value was -0.3VDC.  In practice the short negative voltage would cause a leakage current on the die and result in spurious outputs on the logic (timer start) lines.

This change puts the -V input to -5VDC so that the input signal is kept above the minimum signal value.