Expanded POST test
A clock failure mode whereby the timers could not be stopped was anticipated.  This test arms the counters but does not start them.  It reads the counts back to make sure that they are still zero.

This test is different from a separate test to make sure they can be stopped once started.