Arduino 3.4.4 Fixed ECHO
The ECHO command needs to work for the WiFi version of the PC program to operate correctly.

Previously the ESP-01 driver would overflow and the results were intermittent.

This patch sends the ECHO strings out to the ESP-01 one line at a time and then waits 100ms for the packet to be transmitted