Arduino 3.4.6 Restructured esp01_send
The esp01_send function works by sending an AT command, followed by the text to be output, followed by a null to trigger the send.

Earlier versions of the software would output a short string and then a null many times for a single command, ex ECHO.  This was not consistent as the ESP01 would periodically lose track of what it was doing and no more text would be output.

A change was made to force the AT command on every string, but this would mean +++ (one second delay) AT for every line output which became extremely time consuming.

This patch replaces esp01_send with a state machine that keeps track of what information is waiting in the buffer and then if the buffer overflows, or a null is sent, the buffer is transmitted and the process restarts.

In order for the ESP01 buffer to be flushed, the software must output an empty string, ex esp01_send(0, channel)