Arduino 3.4.11 Adding a factory initialization function
The previous versions of software used init_nonvol to both factory init and user_init the board.  This became confusing particularly with the serial number.

This commit creates a factory_nonvol function that clears up the entire nonvol, while init_nonvol now just touches the things that the user has access to.