Arduino 3.5.1 Refactoring Memory
It turns out that the instruction Serial.print("Hello World"); keeps a copy of "Hello World" in the text (code) segment and makes a copy in the data segment.  Thus Serial.print() consumes RAM space for no good reason.

This patch adds the macro T() to the strings so that the strings are left in the text segment and not consuming RAM.  The RAM requirements went from 78% full to 40% full.

The other change made in this patch was to remove the persistent storage of json_test_number.  The original had the side effect that some self tests could go on forever and not be terminated.  With the addition of the free-form text box in the PC program, persistence across power cycles is not needed any more.