Arduino 3.8.2 Release Candidate 1
There was a situation with the WiFI that if the WiFI connection dropped neither the Arduino or PC would detect the failure.  On the next WiFi output, the transmit would fail, but the Arduino would attempt to transmit for a very long time.

This patch simplifies the esp01_send function and shuts down the channel if not founc