Bug fix in 4.2.7
Inexplicably changing a text from miss:0 to miss:0 caused the program to output garbage.  Investigation showed that the RAM requireements went up by 30 bytes.  This probably caused the stack space to be reduced and stack overflow later on (I'm guessing)

The send_miss function was revised to reduce the amount of RAM and the problem seems to have gone away.