V0.9.2 DAC Revisions
The old DAC driver would write the VREF settings in two operations and sometimes the second value was not written.

This patch changes the DAC write to be a block write of all the settings at once.

The DAC_test was also revised to show a difference between the two channels