ESP32 5.2.6  Fixes for rapid fire
Rapid fire is in two parts:

The regular state machine to compute impacts,
A second state machine to turn the lights on and off.

Between the two of them are a number of semephores, shot_in and shot_out along with rapid_state.  While int the RAPID_ON state, if there is a change in shot_n, then the hardware is reset and continues until the state changes to RAPID_SEND, at which point the accumulated shots are computed and sent.

The DLT log levels were reorganized as:

DLT_DIAG - Hardware related
DLT_APPLICATION - Relates to freetarget.c and compute_hit.c
DLT_DEBUG - General debugging not likely to be used more than once