ESP32 5.2.22 Compressing ECHO command
The echo command was getting long and would no longer fit on a single screen.

The change was to double up the lines so that one line displayed two parameters.

Also, the JSON table was fixed up to remove poniters to double, and replace it with pointers to ints followed by casts to couble (couble).  This was done to save on memory that was not really being used.