Fixed conection bug for BlueTooth
The PC program needs to see something come back from the target to know if it is connected.

Normally connecting to the target over TCPIP generates a signal that results in a message back to the PC.

In the case of BT, there is no connection signal, and no way to generate a reply.

To get around this, the PC sendds * when connected which is a trigger to send a reply